Converted over to using SQL Server and EF
SQLLite was too much of a pain to use and wasn't performing like I
wanted it to when there ended up being over a few gigs of data in a
single table and I needed to do large aggregations. As a result I've
switched over to using SQL Server Express although it could easily use
SQL Server.

To get a new database, the EF Code First model will automatically create
the database if it doesn't already exist when the service or the website
is run.