Page load times fixed
Once the database has a month of data in it, the page load time was
taking two minutes on my server. It was due to fetching all the polled
values at once. It is better to let SQL Server do the number crunching
and then just pull out the bits needed.